+++ common/src/main/java/com/faboslav/friendsandfoes/common/client/render/entity/renderer/PlayerIllusionEntityRenderer.java
@@ -6,13 +6,19 @@ import com.faboslav.friendsandfoes.common.entity.PlayerIllusionEntity;
 import com.faboslav.friendsandfoes.common.util.PlayerSkinProvider;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.model.HumanoidArmorModel;
+import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.geom.ModelLayers;
+import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.client.renderer.entity.HumanoidMobRenderer;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.client.renderer.entity.layers.*;
 
+import net.minecraft.client.renderer.entity.player.PlayerRenderer;
+import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.numbers.StyledFormat;
 import net.minecraft.resources.ResourceLocation;
 
 import java.util.UUID;
@@ -20,6 +26,15 @@ import java.util.UUID;
 //? >=1.21.3 {
 import net.minecraft.client.renderer.entity.state.PlayerRenderState;
 import com.faboslav.friendsandfoes.common.client.render.entity.state.PlayerIllusionRenderState;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.HumanoidArm;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.player.PlayerModelPart;
+import net.minecraft.world.item.*;
+import net.minecraft.world.scores.DisplaySlot;
+import net.minecraft.world.scores.Objective;
+import net.minecraft.world.scores.ReadOnlyScoreInfo;
+import net.minecraft.world.scores.Scoreboard;
 //?}
 
 @SuppressWarnings({"rawtypes", "unchecked"})
@@ -78,6 +93,74 @@ public final class PlayerIllusionEntityRenderer extends LivingEntityRenderer<Pla
 		float partialTick
 	) {
 		super.extractRenderState(playerIllusion, playerIllusionRenderState, partialTick);
+
+		HumanoidMobRenderer.extractHumanoidRenderState(playerIllusion, playerIllusionRenderState, partialTick, this.itemModelResolver);
+		playerIllusionRenderState.leftArmPose = getArmPose(playerIllusion, HumanoidArm.LEFT);
+		playerIllusionRenderState.rightArmPose = getArmPose(playerIllusion, HumanoidArm.RIGHT);
+		playerIllusionRenderState.skin = PlayerSkinProvider.getSkinTextures(playerIllusion);
+		playerIllusionRenderState.arrowCount = playerIllusion.getArrowCount();
+		playerIllusionRenderState.stingerCount = playerIllusion.getStingerCount();
+		playerIllusionRenderState.useItemRemainingTicks = playerIllusion.getUseItemRemainingTicks();
+		playerIllusionRenderState.showHat = playerIllusion.isPartVisible(PlayerModelPart.HAT);
+		playerIllusionRenderState.showJacket = playerIllusion.isPartVisible(PlayerModelPart.JACKET);
+		playerIllusionRenderState.showLeftPants = playerIllusion.isPartVisible(PlayerModelPart.LEFT_PANTS_LEG);
+		playerIllusionRenderState.showRightPants =playerIllusion.isPartVisible(PlayerModelPart.RIGHT_PANTS_LEG);
+		playerIllusionRenderState.showLeftSleeve = playerIllusion.isPartVisible(PlayerModelPart.LEFT_SLEEVE);
+		playerIllusionRenderState.showRightSleeve = playerIllusion.isPartVisible(PlayerModelPart.RIGHT_SLEEVE);
+		playerIllusionRenderState.showCape = playerIllusion.isPartVisible(PlayerModelPart.CAPE);
+	}
+
+	private static HumanoidModel.ArmPose getArmPose(PlayerIllusionEntity playerIllusion, HumanoidArm humanoidArm) {
+		ItemStack itemStack = playerIllusion.getItemInHand(InteractionHand.MAIN_HAND);
+		ItemStack itemStack2 = playerIllusion.getItemInHand(InteractionHand.OFF_HAND);
+		HumanoidModel.ArmPose armPose = getArmPose(playerIllusion, itemStack, InteractionHand.MAIN_HAND);
+		HumanoidModel.ArmPose armPose2 = getArmPose(playerIllusion, itemStack2, InteractionHand.OFF_HAND);
+		if (armPose.isTwoHanded()) {
+			armPose2 = itemStack2.isEmpty() ? HumanoidModel.ArmPose.EMPTY : HumanoidModel.ArmPose.ITEM;
+		}
+
+		return playerIllusion.getMainArm() == humanoidArm ? armPose : armPose2;
+	}
+
+	private static HumanoidModel.ArmPose getArmPose(PlayerIllusionEntity playerIllusion, ItemStack itemStack, InteractionHand interactionHand) {
+		if (itemStack.isEmpty()) {
+			return HumanoidModel.ArmPose.EMPTY;
+		} else if (!playerIllusion.swinging && itemStack.is(Items.CROSSBOW) && CrossbowItem.isCharged(itemStack)) {
+			return HumanoidModel.ArmPose.CROSSBOW_HOLD;
+		} else {
+			if (playerIllusion.getUsedItemHand() == interactionHand && playerIllusion.getUseItemRemainingTicks() > 0) {
+				ItemUseAnimation itemUseAnimation = itemStack.getUseAnimation();
+				if (itemUseAnimation == ItemUseAnimation.BLOCK) {
+					return HumanoidModel.ArmPose.BLOCK;
+				}
+
+				if (itemUseAnimation == ItemUseAnimation.BOW) {
+					return HumanoidModel.ArmPose.BOW_AND_ARROW;
+				}
+
+				if (itemUseAnimation == ItemUseAnimation.SPEAR) {
+					return HumanoidModel.ArmPose.THROW_SPEAR;
+				}
+
+				if (itemUseAnimation == ItemUseAnimation.CROSSBOW) {
+					return HumanoidModel.ArmPose.CROSSBOW_CHARGE;
+				}
+
+				if (itemUseAnimation == ItemUseAnimation.SPYGLASS) {
+					return HumanoidModel.ArmPose.SPYGLASS;
+				}
+
+				if (itemUseAnimation == ItemUseAnimation.TOOT_HORN) {
+					return HumanoidModel.ArmPose.TOOT_HORN;
+				}
+
+				if (itemUseAnimation == ItemUseAnimation.BRUSH) {
+					return HumanoidModel.ArmPose.BRUSH;
+				}
+			}
+
+			return HumanoidModel.ArmPose.ITEM;
+		}
 	}
 	//?} else {
 	//?}
@@ -101,13 +184,7 @@ public final class PlayerIllusionEntityRenderer extends LivingEntityRenderer<Pla
 		//? >=1.21.3 {
 		return playerIllusionRenderState.skin.texture();
 		//?} else {
-		/*UUID uuid = playerIllusion.getPlayerUuid();
-
-		if (uuid == null) {
-			uuid = playerIllusion.getUUID();
-		}
-
-		return PlayerSkinProvider.getSkinTextures(uuid).texture();
+		/*return PlayerSkinProvider.getSkinTextures(playerIllusion).texture();
 		*///?}
 	}
 }